* Compilation fixes for Cygwin.
git-svn-id: svn://svn.minisip.org/minisip/trunk@1008 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ minisip/p2t/GroupListServer.cxx
@@ -56,7 +56,7 @@ GroupListServer::~GroupListServer(){
 void GroupListServer::run(){
 	
 #ifdef DEBUG_OUTPUT
-	mdbg << "GroupListServer:: Accepting connections at " << ip<<":"<<port<<end;
+	mdbg << "GroupListServer:: Accepting connections at " << ip<<":"<<(int)port<<end;
 #endif
 	
 	TCPSocket* socket;
@@ -145,7 +145,7 @@ void GroupListServer::stop(){
 	
 	
 #ifdef DEBUG_OUTPUT
-	mdbg << "GroupListServer:: "<< ip<<":"<<port<<" stopped!"<<end;
+	mdbg << "GroupListServer:: "<< ip<<":"<<(int)port<<" stopped!"<<end;
 #endif
 }
